requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.23.0